Include <climits> for UINT_MAX + remove a cast, unsigned can NOT be greater UINT_MAX.

+++ hist/spectrumpainter/src/TSpectrum2Painter.cxx
@@ -25,6 +25,7 @@
 // algorithms for analysis of multidimensional experimental nuclear data. Acta
 // Pysica Slovaca Vol. 54/ 4 (2004), pp. 385-400.
 
+#include <climits>
 
 #include "TROOT.h"
 #include "TClass.h"
@@ -485,8 +486,8 @@ void TSpectrum2Painter::ColorModel(unsigned ui, unsigned ui1, unsigned ui2,
       a     = fRainbow1Step;
       a     = ui*a;
       a     = ui1+a;
-      if ((unsigned)a>=UINT_MAX) uinc1 = UINT_MAX;
-      else                       uinc1 = (unsigned)a;
+      if (a >= UINT_MAX) uinc1 = UINT_MAX;
+      else               uinc1 = (unsigned)a;
       upom  = uinc1%256;
       i     = (uinc1-upom)/256;
       if ((i%2)==0) uinc1 = upom;
